+++ assets/src/components/nav/topNav.tsx
@@ -1,10 +1,9 @@
-import React, { useContext } from "react"
+import React from "react"
 import { Link } from "react-router-dom"
 import { logoIcon, refreshIcon } from "../../helpers/icon"
 import { reload } from "../../models/browser"
 
 const TopNav = (): JSX.Element => {
-
   return (
     <div className="m-top-nav">
       <Link className="m-top-nav__logo" to="/" title="Skate">
